+++ pkg/install/deployment.go
@@ -25,7 +25,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/vmware-tanzu/velero/internal/velero"
+	images "github.com/vmware-tanzu/velero/internal/velero/images"
 	"github.com/vmware-tanzu/velero/pkg/builder"
 	"github.com/vmware-tanzu/velero/pkg/util/kube"
 )
@@ -195,7 +195,7 @@ func WithKeepLatestMaintenanceJobs(keepLatestMaintenanceJobs int) podTemplateOpt
 func Deployment(namespace string, opts ...podTemplateOption) *appsv1.Deployment {
 	// TODO: Add support for server args
 	c := &podTemplateConfig{
-		image: velero.DefaultVeleroImage(),
+		image: images.DefaultVeleroImage(),
 	}
 
 	for _, opt := range opts {
