+++ pkg/restore/actions/pod_volume_restore_action_test.go
@@ -28,7 +28,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/kubernetes/fake"
 
-	veleroimage "github.com/vmware-tanzu/velero/internal/velero"
+	veleroimage "github.com/vmware-tanzu/velero/internal/velero/images"
 	velerov1api "github.com/vmware-tanzu/velero/pkg/apis/velero/v1"
 	"github.com/vmware-tanzu/velero/pkg/builder"
 	"github.com/vmware-tanzu/velero/pkg/buildinfo"
